add WindowProxy and change float types to integer types

+++ examples/window_params.rs
@@ -10,7 +10,7 @@ impl Application {
 }
 
 impl wita::EventHandler for Application {
-    fn resized(&mut self, _: &wita::Window, size: wita::PhysicalSize<f32>) {
+    fn resized(&mut self, _: &wita::Window, size: wita::PhysicalSize<u32>) {
         println!("resized: {:?}", size);
     }
 
@@ -26,7 +26,7 @@ impl wita::EventHandler for Application {
                 }
                 wita::VirtualKey::Char('R') => window.set_position(wita::ScreenPosition::new(0, 0)),
                 wita::VirtualKey::Char('S') => {
-                    window.set_inner_size(wita::LogicalSize::new(256.0, 256.0))
+                    window.set_inner_size(wita::LogicalSize::new(256, 256))
                 }
                 _ => (),
             }
