+++ src/device.rs
@@ -20,7 +20,7 @@ pub enum MouseButton {
 /// A mouse cursor position and pressed mouse buttons.
 #[derive(Clone, Debug)]
 pub struct MouseState<'a> {
-    pub position: PhysicalPosition<f32>,
+    pub position: PhysicalPosition<i32>,
     pub buttons: &'a [MouseButton],
 }
